requirements.txt
streamlit
pandas
openpyxl
reportlab
langgraph
requests
python-dotenv
google-generativeai